requirements.txt
django==5.2.4
django-encrypted-model-fields==0.6.5
dotenv==1.1.1 
psycopg==3.2.9
psycopg-binary==3.2.9
djangorestframework==3.16.0